+++ glibc/pthread/samples/create-deadline.c
@@ -24,7 +24,7 @@ static volatile int done;
 
 void *run_deadline(void *data)
 {
-	struct __sched_attr attr;
+	struct sched_attr attr;
 	int x = 0;
 	int ret;
 	unsigned int flags = 0;