+++ syscall/samples/sched/sched_helpers.c
@@ -148,13 +148,13 @@ long int sys_getcpu(unsigned *cpu, unsigned *node)
 	return syscall(__NR_getcpu, cpu, node);
 }
 
-int sys_sched_setattr(pid_t pid, const struct __sched_attr *attr,
+int sys_sched_setattr(pid_t pid, const struct sched_attr *attr,
 		      unsigned int flags)
 {
 	return syscall(__NR_sched_setattr, pid, attr, flags);
 }
 
-int sys_sched_getattr(pid_t pid, struct __sched_attr *attr, unsigned int size,
+int sys_sched_getattr(pid_t pid, struct sched_attr *attr, unsigned int size,
 		      unsigned int flags)
 {
 	return syscall(__NR_sched_getattr, pid, attr, size, flags);